refactor(tests): extract shared encounter form helpers
